`Cache` class in separate file
Introduce Metica.SDK.Caching namespace. Metica.Unity still too broad and for now must still be included (using).
