+++ SDK/Runtime/Caching/Cache.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9c17b876187dcf9469008337b8693dfc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
