+++ SDK/Runtime/RemoteConfigManager.cs
@@ -2,7 +2,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Newtonsoft.Json;
 
 namespace Metica.Unity
 {
